Rename quiz components
Refs #40
